# Refactoring smoothing/averaging of sensor values.