--- a/app/src/main/res/values/dimens.xml
+++ b/app/src/main/res/values/dimens.xml
@@ -4,9 +4,9 @@
     <dimen name="level_border_width">5dip</dimen>
     <dimen name="level_border_height">0dip</dimen>
     <!-- Display -->
-    <dimen name="display_padding">8dip</dimen>
+    <dimen name="display_padding">6dip</dimen>
     <dimen name="display_border">1dip</dimen>
-    <dimen name="display_gap">20dip</dimen>
+    <dimen name="display_gap">10dip</dimen>
     <!-- Sensor -->
     <dimen name="sensor_gap">5dip</dimen>
     <!-- Marker -->
@@ -23,6 +23,5 @@
     <dimen name="lock_text">32dip</dimen>
     <!-- INFO -->
     <dimen name="info_text">12dip</dimen>
-    <dimen name="arrow_gap">10dip</dimen>
-    <dimen name="arrow_height">24dip</dimen>
+    <dimen name="arrow_padding">10dip</dimen>
 </resources>