--- a/fastlane/metadata/android/de-DE/changelogs/19.txt
+++ b/fastlane/metadata/android/de-DE/changelogs/19.txt
@@ -1,0 +1,1 @@
+GUI verbessert