… fastlane/metadata/android/en-US/images/phoneScreenshots/Bubble1.png
… fastlane/metadata/android/en-US/images/phoneScreenshots/Bubble2.png