An exponential smoothing algorithm based on event timestamps has been added to OrientationProvider.

It replaces the viscosity emulating code in LevelPainter. And the heuristic code to calm down number displays is not needed any more.

Averaged values are now used for calibration rather than taking a single data point.

It will also have a calming effect on the number display