@woheller69 2023-01-15
Nice idea, but the bubble behaves quite strange when the device is flat on the table.
The bubble behaves as if it was repelled by something