@lufebe16 2023-01-16
This is a problem that exists also in the current published version. Only the dynamics have changed. So it is more obvious.

The real problem is located in the code at the end of onOrientationChanged in LevelPainter. There is a 'correction des angles a plat - la bulle ne doit pas sortir du niveau'. The calculations are incorrect. They affect the angle in x (or y) when angle y (or x) is near zero. The digit display is not affected.

If that code is removed the problem disappears. But the bubble can go out of the circle, if not one of the angles is zero. So we would need a replacement.