@woheller69 2023-01-16
onOrientationChanged is only called when orientation is changed e.g. from Bottom to Landing.
But while we are in Landing mode this is not called anymore, so it cannot influence the behaviour of the bubble.