@woheller69 2023-01-17
Thanks, but I am still not convinced. I tested your fork on several devices and on older phones the movement of the bubble is not smooth. The bubble is moving in steps.
But I removed the stuff from onOrientationChanged and switched to your algo for keeping the bubble in the circle.