@lufebe16 2023-01-18
Thanks for the new arrows. It makes the look and feel more consistent over screens. And it adds the originally missing sign information to landing mode.

The problem of the bubble moving in steps could come from the eco mode (i.e. limiting the calculation to the framerate) introduced in onSensorChanged at the beginning. There is
`if (deltaTime < framedist) return;
`
Maybe it helps to remove that line.

But yes I can reproduce it on a samsung tablet  (about 10 years old) with a 32 bit equinox processor (and maybe soft float emulation!). So it might be the Math.pow call being to heavy for it.  Removing the eco mode here makes it even worse.