@lufebe16 2023-01-19
Here just my results of some timing tests, tested on an older device (10 years 32 bit) and a phone (5 years 64 bit) .

The pow function is out of scope. Also on the older device it needs 0 ms for dozens of measurments, before I see a single 1.

onSensorChanged delivers values with event-timestamps at a regular basis as expected. Values are 67 and 40 ms for the two devices.

Method onSensorChanged itself is called very irregularily on both devices. The difference between calls ranges from 5 to 120 ms (!). So that adds a big jitter to the updates in LevelPainter. Especially a 100 ms delay will be recognizable if applied directly to the screen graphics. An additional smoothing at the beginning of the draw process (~ 100 ms) is therefore still required.

The timing of the paint routine on both devices is equal and varies from 25 to 40 ms. So it does far not reach its limitation by 16 ms (frameRate).