@woheller69 2023-01-19
I have done some other tests and moved my simple (70/30) smoothing algorithm to orientation provider. 
That works but  causes other issues.
E.g. when locking the orientation the bubble does not move to the end anymore and even moves back when the angle is close to 90°.
So I tend to keep it as it is. When I revived the app I did not have the intention to reinvent it...