@woheller69 2023-01-20
Thanks for your proposals. There are also many things I would do differently in this app but I did not want to completely rewrite it.
I simply wanted to revive it as it had not been updated since 2014...